+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#788](https://github.com/alleslabs/celatone-frontend/pull/788) Add voting power chart component
 - [#763](https://github.com/alleslabs/celatone-frontend/pull/763) Proposal status summary top
 - [#765](https://github.com/alleslabs/celatone-frontend/pull/765) Support Proposal Cancelled status
 - [#760](https://github.com/alleslabs/celatone-frontend/pull/760) Add proposal messages